some internal Skill API refactoring
