+++ src/main/java/com/github/manasmods/manascore/client/skill/ClientSkillEventListenerHandler.java
@@ -1,4 +1,4 @@
-package com.github.manasmods.manascore.skill;
+package com.github.manasmods.manascore.client.skill;
 
 import com.github.manasmods.manascore.ManasCore;
 import com.github.manasmods.manascore.api.skills.ManasSkillInstance;
@@ -6,12 +6,13 @@ import com.github.manasmods.manascore.api.skills.SkillAPI;
 import com.github.manasmods.manascore.api.skills.listener.KeyInputListener;
 import net.minecraft.client.Minecraft;
 import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.client.event.EntityViewRenderEvent;
 import net.minecraftforge.client.event.InputEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 
 @Mod.EventBusSubscriber(modid = ManasCore.MOD_ID, bus = Mod.EventBusSubscriber.Bus.FORGE, value = Dist.CLIENT)
-class ClientEventListenerHandler {
+class ClientSkillEventListenerHandler {
     private static final Minecraft mc = Minecraft.getInstance();
 
     @SubscribeEvent
@@ -23,5 +24,10 @@ class ClientEventListenerHandler {
                 listener.onKeyInput(instance, e);
             }
         }
+    }
+
+    @SubscribeEvent
+    public static void onRenderEntity(final EntityViewRenderEvent.CameraSetup e){
+
     }
 }
